+++ R/trace.R
@@ -397,17 +397,28 @@ c.rlang_trace <- function(...) {
 #' @export
 format.rlang_trace <- function(x,
                                ...,
-                               simplify = c("none", "collapse", "branch"),
+                               simplify = c("none", "branch"),
                                max_frames = NULL,
                                dir = getwd(),
                                srcrefs = NULL,
                                drop = FALSE) {
   switch(
-    arg_match(simplify),
+    arg_match_simplify(simplify),
     none = trace_format(x, max_frames, dir, srcrefs, drop = drop),
-    collapse = trace_format_collapse(x, max_frames, dir, srcrefs),
     branch = trace_format_branch(x, max_frames, dir, srcrefs)
   )
+}
+
+arg_match_simplify <- function(simplify, call = caller_env()) {
+  if (is_string(simplify, "collapse")) {
+    deprecate_collapse()
+    simplify <- "none"
+  }
+  arg_match0(simplify, c("none", "branch"), error_call = call)
+}
+
+deprecate_collapse <- function() {
+  warn_deprecated("`\"collapse\"` is deprecated as of rlang 1.1.0.\nPlease use `\"none\"` instead.")
 }
 
 trace_format <- function(trace, max_frames, dir, srcrefs, drop = FALSE) {
@@ -427,11 +438,6 @@ trace_format <- function(trace, max_frames, dir, srcrefs, drop = FALSE) {
   cli_tree(tree)
 }
 
-trace_format_collapse <- function(trace, max_frames, dir, srcrefs) {
-  trace <- trace_simplify_collapse(trace)
-  trace_format(trace, max_frames, dir, srcrefs, drop = TRUE)
-}
-
 trace_format_branch <- function(trace, max_frames, dir, srcrefs) {
   trace <- trace_simplify_branch(trace)
   tree <- trace_as_tree(trace, dir = dir, srcrefs = srcrefs)
@@ -440,18 +446,6 @@ trace_format_branch <- function(trace, max_frames, dir, srcrefs) {
   tree <- vec_slice(tree, -1)
 
   cli_branch(tree, max = max_frames)
-}
-
-format_collapsed <- function(what, n) {
-  if (n > 0L) {
-    call_text <- pluralise_n(n, "call", "calls")
-    n_text <- sprintf(" with %d more %s", n, call_text)
-    n_text <- silver(n_text)
-  } else {
-    n_text <- ""
-  }
-
-  paste0(what, n_text)
 }
 
 cli_branch <- function(tree,
@@ -531,11 +525,11 @@ zip_chr <- function(xs, ys) {
 #' @export
 print.rlang_trace <- function(x,
                               ...,
-                              simplify = c("none", "branch", "collapse"),
+                              simplify = c("none", "branch"),
                               max_frames = NULL,
                               dir = getwd(),
                               srcrefs = NULL) {
-  simplify <- arg_match(simplify)
+  simplify <- arg_match_simplify(simplify)
   cat_line(format(x, ...,
     simplify = simplify,
     max_frames = max_frames,
@@ -587,39 +581,6 @@ trace_trim_env_idx <- function(n, frames, to) {
   seq2(start, n)
 }
 
-set_trace_skipped <- function(trace, id, n) {
-  trace$collapsed[[id]] <- n
-  trace
-}
-set_trace_collapsed <- function(trace, id, n) {
-  trace$collapsed[[id - n]] <- n
-  trace
-}
-n_collapsed <- function(trace, id) {
-  call <- trace$call[[id]]
-
-  if (is_eval_call(call)) {
-    # When error occurs inside eval()'s frame at top level, there
-    # might be only one frame and nothing to collapse
-    if (id > 1L && is_eval_call(trace$call[[id - 1L]])) {
-      n <- 1L
-    } else {
-      n <- 0L
-    }
-    return(n)
-  }
-
-  if (identical(call, quote(function_list[[i]](value)))) {
-    return(6L)
-  }
-
-  if (identical(call, quote(function_list[[k]](value)))) {
-    return(7L)
-  }
-
-  0L
-}
-
 is_eval_call <- function(call) {
   is_call2(call, c("eval", "evalq"), ns = c("", "base"))
 }
@@ -641,17 +602,7 @@ trace_simplify_branch <- function(trace) {
     id <- 0L
   }
 
-  trace$collapsed <- 0L
-
   while (id != 0L) {
-    n_collapsed <- n_collapsed(trace, id)
-
-    if (n_collapsed) {
-      trace <- set_trace_collapsed(trace, id, n_collapsed)
-      next_id <- id - n_collapsed
-      id <- next_id
-    }
-
     # Set `old_visible` to avoid uninformative calls in position 1 to
     # be included (see below)
     if (is_uninformative_call(trace$call[[id]])) {
@@ -719,68 +670,11 @@ is_winch_frame <- function(call) {
   grepl("^[/\\\\].+[.]", name)
 }
 
-trace_simplify_collapse <- function(trace) {
-  if (!trace_length(trace)) {
-    return(trace)
-  }
-
-  parents <- trace$parent
-
-  old_visible <- trace$visible
-  visible <- rep_along(old_visible, FALSE)
-
-  id <- trace_length(trace)
-
-  trace$collapsed <- 0L
-
-  while (id > 0L) {
-    n_collapsed <- n_collapsed(trace, id)
-
-    if (n_collapsed) {
-      trace <- set_trace_collapsed(trace, id, n_collapsed)
-      next_id <- id - n_collapsed
-
-      # Rechain child of collapsed parent to correct parent
-      parents[[id + 1L]] <- next_id
-
-      id <- next_id
-    }
-
-    visible[[id]] <- TRUE
-    parent_id <- parents[[id]]
-    id <- dec(id)
-
-    # Collapse intervening call branches
-    n_skipped <- 0L
-    while (id != parent_id) {
-      sibling_parent_id <- parents[[id]]
-
-      if (sibling_parent_id == parent_id) {
-        trace <- set_trace_skipped(trace, id, n_skipped)
-        visible[[id]] <- TRUE
-        n_skipped <- 0L
-      } else {
-        n_skipped <- inc(n_skipped)
-      }
-
-      id <- dec(id)
-    }
-  }
-
-  trace$visible <- visible & old_visible
-  trace$parent <- parents
-
-  trace
-}
-
 
 # Printing ----------------------------------------------------------------
 
 trace_as_tree <- function(trace, dir = getwd(), srcrefs = NULL, drop = FALSE) {
-  if (is_null(trace$collapsed)) {
-    trace$collapsed <- vec_recycle(0L, trace_length(trace))
-  }
-  call_text_data <- trace[c("call", "collapsed", "namespace", "scope")]
+  call_text_data <- trace[c("call", "namespace", "scope")]
   call_text <- chr(!!!pmap(call_text_data, trace_call_text))
 
   srcrefs <- srcrefs %||% peek_option("rlang_trace_format_srcrefs") %||% TRUE
@@ -870,11 +764,7 @@ node_type <- function(ns, children) {
 }
 
 # FIXME: Add something like call_deparse_line()
-trace_call_text <- function(call, collapsed, namespace, scope) {
-  if (collapsed && length(call) > 1L) {
-    call <- call2(call[[1]], quote(...))
-  }
-
+trace_call_text <- function(call, namespace, scope) {
   if (is_call(call) && is_symbol(call[[1]])) {
     if (scope %in% c("::", ":::") && !is_na(namespace)) {
       call[[1]] <- call(scope, sym(namespace), call[[1]])
@@ -887,11 +777,6 @@ trace_call_text <- function(call, collapsed, namespace, scope) {
     text <- paste0("global ", text)
   } else if (is_string(scope, "local") && !is_na(namespace)) {
     text <- paste0(namespace, " (local) ", text)
-  }
-
-  if (collapsed) {
-    n_collapsed_text <- sprintf(" ... +%d", collapsed)
-    text <- format_collapsed(paste0("[ ", text, " ]"), collapsed)
   }
 
   text